fix: fixes issues with stories and UserProfile after moving icons
